+++ build-logic/convention/src/main/java/com/bd/convention/android-library-convention.gradle.kts
@@ -1,5 +1,4 @@
 import com.bd.convention.BuildConstants
-import java.util.Properties
 
 plugins {
     id("com.android.library")
@@ -18,12 +17,6 @@ android {
 
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
         consumerProguardFiles("consumer-rules.pro")
-
-        val keystoreFile = project.rootProject.file("local.properties")
-        val properties = Properties()
-        properties.load(keystoreFile.inputStream())
-        val apiKey = properties.getProperty("TMDB_API_KEY") ?: ""
-        buildConfigField(type = "String", name = "TMDB_API_KEY", value = apiKey)
     }
 
     buildTypes {
